refactor: move execute_compaction to compaction module
Compaction execution is compaction logic, not agent logic. Agent now
imports and calls it.
